+++ GameSpeed2U/src/item/console/ConsoleHashMap.java
@@ -1,4 +1,4 @@
-package item;
+package item.console;
 
 import java.util.HashMap;
