+++ GameSpeed2U/src/item/console/ConsoleList.java
@@ -1,4 +1,4 @@
-package item;
+package item.console;
 
 import java.io.IOException;
 import java.io.PrintWriter;
@@ -10,6 +10,8 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+
+import item.GenerateItemHtmlHandler;
 import main.*;
 
 @WebServlet(name = "ConsoleList", urlPatterns = "/ConsoleList")
