+++ GameSpeed2U/src/item/game/Game.java
@@ -1,4 +1,6 @@
-package item;
+package item.game;
+
+import item.CatalogItem;
 
 public class Game extends CatalogItem {
