+++ GameSpeed2U/src/item/game/GameHashMap.java
@@ -1,4 +1,6 @@
-package item;
+package item.game;
+
+import item.game.Game;
 
 import java.util.HashMap;
