+++ GameSpeed2U/src/item/tablet/Tablet.java
@@ -1,4 +1,4 @@
-package item;
+package item.tablet;
 
 public class Tablet {
 	private String name;
